C114 Stable point, add react scan
