+++ .husky/pre-commit
@@ -1,2 +1,1 @@
 yarn prettier . --write
-yarn lint
