+++ src/components/BaseMap/BaseMap.tsx
@@ -25,7 +25,7 @@ import { cogProtocol } from '@geomatico/maplibre-cog-protocol'
 import { LayerInfo } from '../../data/mapData'
 import useResponsive from '../../hooks/useResponsive'
 
-import { updateChartData } from '../../utils/chartUtils'
+import { updateChartData, loadChartData } from '../../utils/chartUtils'
 import LoadingState from '../LoadingState/LoadingState'
 import { RegionOption } from '../../types/RegionDataTypes'
 import {
@@ -62,7 +62,7 @@ export default function BaseMap({
   const defaultMapZoom = 10
   const mapRef = useRef<MapRef | null>(null)
   // const currentLngLat = useMemo(() => [defaultLng, defaultLat], [defaultLng, defaultLat])
-  const [currentLngLat] = useState<[number, number]>([defaultLng, defaultLat])
+  // const [, setCurrentLngLat] = useState<[number, number]>([defaultLng, defaultLat])
   const [layerErrors, setLayerErrors] = useState<Record<string, string>>({})
   const polygonHoverRef = useRef<string | number | null>(null)
   const polygonClickRef = useRef<string | number | null>(null)
@@ -104,40 +104,45 @@ export default function BaseMap({
     map?.jumpTo(jumpOptions)
   }, [selectedRegion])
 
-  const loadChartData = useCallback(
-    (point, activeLayers) => {
-      if (!mapRef.current) {
-        return
-      }
-      const map = mapRef.current?.getMap()
-      const zoom = mapRef.current?.getMap().getZoom()
-      //doesn't account for if the layer hasn't loaded yet
-      if (activeLayers.length > 0) {
-        //query the layers corresponding with charts and with layers that are on
-        const features = map.queryRenderedFeatures(point, {
-          layers: activeLayers,
-        })
-
-        if (features.length > 0) {
-          const tempSkipLayers = ['global']
-          //only grab the topmost data layer to pull point from
-          const firstFeature = features[0]
-
-          //TEMP: remove tempSkipLayers when all region data is available
-          if (!tempSkipLayers.includes(firstFeature.layer.id)) {
-            const mappedRegion = mapRegionSelected(firstFeature, currentLngLat, zoom)
-            setSelectedRegion(mappedRegion)
-
-            //trigger chart data update
-            updateChartData(firstFeature, setChartConfigData)
-          } else {
-            setChartConfigData(null)
-          }
-        }
-      }
-    },
-    [currentLngLat, setChartConfigData, setSelectedRegion],
-  )
+  // const loadChartData = useCallback(
+  //   (point, activeLayers, givenMap?) => {
+  //     let map
+  //     if (!givenMap) {
+  //       if (!mapRef.current) {
+  //         return
+  //       }
+  //       map = mapRef.current?.getMap()
+  //     } else {
+  //       map = givenMap
+  //     }
+  //
+  //     //doesn't account for if the layer hasn't loaded yet
+  //     if (activeLayers.length > 0) {
+  //       //query the layers corresponding with charts and with layers that are on
+  //       const features = map.queryRenderedFeatures(point, {
+  //         layers: activeLayers,
+  //       })
+  //
+  //       if (features.length > 0) {
+  //         const tempSkipLayers = ['global']
+  //         //only grab the topmost data layer to pull point from
+  //         const firstFeature = features[0]
+  //
+  //         //TEMP: remove tempSkipLayers when all region data is available
+  //         if (!tempSkipLayers.includes(firstFeature.layer.id)) {
+  //           const mappedRegion = mapRegionSelected(firstFeature)
+  //           setSelectedRegion(mappedRegion)
+  //
+  //           //trigger chart data update
+  //           updateChartData(firstFeature, setChartConfigData)
+  //         } else {
+  //           setChartConfigData(null)
+  //         }
+  //       }
+  //     }
+  //   },
+  //   [setChartConfigData, setSelectedRegion],
+  // )
 
   if (
     !import.meta.env.VITE_MAPTILER_API_KEY ||
@@ -199,13 +204,14 @@ export default function BaseMap({
 
   const handleMapLoad = () => {
     setIsMapLoaded(true)
-    const activeLayers = getActiveLayers(mapLayers)
-    loadChartData(defaultPoint, activeLayers)
 
     const map = mapRef.current?.getMap()
     if (!map) {
       return
     }
+    //unnecessary while loading mock data to start
+    // const activeLayers = getActiveLayers(mapLayers)
+    // loadChartData(defaultPoint, activeLayers, map)
 
     const watershedLayer = mapLayers.find((l) => l.layerId === 'watershed')
 
@@ -225,8 +231,8 @@ export default function BaseMap({
     const clickBound = (e) => {
       polygonClickHandler(map, e, watershedLayer)
       // TODO: find correct location for these w/o triggering an entire app re-render
-      // const activeLayers = getActiveLayers(mapLayers)
-      // loadChartData(e.point, activeLayers)
+      const activeLayers = getActiveLayers(mapLayers)
+      loadChartData(e.point, activeLayers, map)
     }
 
     polygonHoverBoundRef.current = hoverBound
@@ -244,18 +250,6 @@ export default function BaseMap({
         promoteId="watershed_id"
         url={`pmtiles://${layer.link}`}
       >
-        <Layer
-          id={layer.layerId}
-          type="fill"
-          key={`${layer.layerId}-${index}`}
-          source={layer.sourceId}
-          source-layer={layer.sourceName}
-          beforeId="label_airport" // label_airport is one of the first label layers, this ensures custom layers appear below all labels
-          paint={{
-            'fill-color': 'rgba(0,0,0,0)',
-            'fill-outline-color': layer.outlineColor,
-          }}
-        />
         <Layer
           id={`${layer.layerId}-lines`}
           type="line"
@@ -276,6 +270,18 @@ export default function BaseMap({
               polygonOutlineSelectColor,
               'rgba(0,0,0,0)',
             ],
+          }}
+        />
+        <Layer
+          id={layer.layerId}
+          type="fill"
+          key={`${layer.layerId}-${index}`}
+          source={layer.sourceId}
+          source-layer={layer.sourceName}
+          beforeId="label_airport" // label_airport is one of the first label layers, this ensures custom layers appear below all labels
+          paint={{
+            'fill-color': 'rgba(0,0,0,0)',
+            'fill-outline-color': layer.outlineColor,
           }}
         />
       </Source>
